Allow providing a executor to be able to run in parallel
Instead of just running one callable at a time it can be
quite useful to run multiple callables that the schedule
says to run in parallel so that it is easily possible to
avoid blocking up the schedule queue.

To make this possible, allow for providing a executor
factory function that can create a executor that will be used
to run callables in parallel (by default it uses the sync
executor, which does *not* run things in parallel to retain
the existing behavior).

Change-Id: I85deacd3ab838beb20ba0a9d014c87e7fd54b806