Allow providing customized event + condition classes/factories
When a green thread pool executor (and eventlet) is used and
the threading module is *not* patched it is required to avoid
using the native threading condition and event classes to ensure
that the eventlet loop is context-switched into.

To make it possible for this periodic worker to work in those
such usages allow for providing the condition and event class
that will be used internally (which should be green equivalents
when eventlet is used).

Change-Id: I82e10e4318936745e84db3418af619523d1cb0c3