Allow providing customized scheduling strategies
Instead of fixing on having a scheduling strategy that
uses the last time the callback started + the callbacks
periodicity as the next time the callback should run allow
for providing a customized callback 'strategy' that can
be used to provide different ways of determining when a
callback should run (given metrics about prior runs).

This adds 4 built-in strategies that can be used to
aid in picking when next runs will happen. It also adds
a enhanced metrics output that happens when the periodic
worker finishes.

Change-Id: I0bee6912830b88de3d75e543689c54fd02801dd5