Ensure we check callback '_is_periodic' on add
Instead of always adding we should make sure we check
the callback attribute '_is_periodic' and avoid adding
if this is set to false.

This always has the `add` function return whether the
add was successful or not (using the '_is_periodic' attribute
to affect this return result).

Whats returned from `add` (and is also accessible now via
`iter_watchers`) is a new watcher object which provides a
read-only view on the callback (and its metrics) as it gets ran
by the worker.

Change-Id: Ie77248f04125c9f61a336d45f25ac1eaa53ab497