Ensure doctesting and doc8 testing done in py27 env
This ensures that sphinx doc tests are ran and validated
and also runs doc8 to ensure style checks are done and adjusts
some of the current examples and code to fix found issues.

The process based executor doctest is disabled as it appears
sphinx creates the code for these examples in a local namespace
which the process based executor can't locate (and therefore it
dies).

Change-Id: I825f3a4b673acd66cc6e05d0d1057ed5c72a759d