Provide a thread differentiation attribute on executors
Have each executor expose a 'threading' attribute on itself
or its class which can be used by executor consumers to be able
to construct locks, rlocks, conditions, and event objects that
are most compatible with the executors execution style.

This allows for creating libraries that can use the executor defined
objects internally, and avoids the need to have to monkey patch all
the things (since typically libraries just stick to the built-in locks
and conditions and such, and therefore require monkey patching to work
correctly under different executor models).

Change-Id: I6c09a7660cf6a89142064d0db31846481ae7d1be