Ensure new entry in immediates gets processed during wait(s)
When a new callback is added, it should cause the waiting thread
that processes callbacks to break out of its while loop and process
that callback.

This fixes that case, making sure that a new immediate causes the
waiting loop to be broken out of and therefore makes sure the new
immediate callback is processed as expected.

Change-Id: Ia590f3d7be47e4045ff40de0e921112d3318d788