Remove green worker reference to its executor
There really is no need to give a green worker
a reference to its creator and it hinders/makes it
harder for python to garbage collect the executor
and its workers.

It is also not used, so let's make the python
garbage collectors life easier by not even passing
it in.

Change-Id: I0f9d27663c3ca5ca908ef6f90f3d1d0f6b5f9104