Some minor refactoring and comment/note addition
In the internal '_run' instead of having a loop that
has various cases with quite a bit of code per case instead
have the loop call into smaller helper functions that handle
the different types of processing the loop does.

Also adds a note to the 'reset' method that explains why
the metrics keys are overwritten vs being copied/replaced
with a new copy.

Change-Id: I5ac1794d5803850120be4e67b85efe0449aa1891