Expose underlying futures.CancelledError
When a future is cancelled and the result is fetched
it will raise a CancelledError which is quite useful to
be able to catch.

So to enable catching that specific exception expose it
in futurist __init__.py so that people can easily catch
it just by catching futurist.CancelledError and handling
the outcome of that.

Change-Id: I7bc1838aadcbb8cd0d635da83bb1e1b1ec761e89