clean up test job configuration
Make the py27 and py35 jobs run the same commands by putting them into
the base testenv block.

Move the doc8 job into the pep8 environment to run with the other
linters.

Change the way we invoke sphinx to run the doctests to avoid issues
with pbr's sphinx integration. This also required removing skipdist
and usedevelop to ensure that all of the pbr-generated files needed by
sphinx exist before we run sphinx.

Change the way sphinx is invoked in the docs environment to avoid
issues with pbr's sphinx integration.

Change-Id: If8a24428eca4512a66dbb361e55cd73aa92c4da8
Signed-off-by: Doug Hellmann <doug@doughellmann.com>